refactor: move IO operations over to separate file
Reduces the amount of manipulation required massively
